+++ src/test/groovy/ru/vyarus/dropwizard/guice/bundles/bootstrap/GBootstrapApplication.groovy
@@ -26,7 +26,7 @@ class GBootstrapApplication extends Application<TestConfiguration> {
 
     @Override
     void initialize(Bootstrap<TestConfiguration> bootstrap) {
-        bootstrap.addBundle(GuiceBundle.<TestConfiguration> builder()
+        bootstrap.addBundle(GuiceBundle.builder()
                 .bundles(
                 new GuiceyBundle() {
                     @Override