+++ src/test/groovy/ru/vyarus/dropwizard/guice/bundles/manual/ManualBundlesApplication.groovy
@@ -18,7 +18,7 @@ class ManualBundlesApplication extends Application<TestConfiguration> {
 
     @Override
     void initialize(Bootstrap<TestConfiguration> bootstrap) {
-        bootstrap.addBundle(GuiceBundle.<TestConfiguration> builder()
+        bootstrap.addBundle(GuiceBundle.builder()
                 .disableBundles(GuiceRestrictedConfigBundle)
                 .extensions(DummyTask, DummyResource, DummyManaged)
                 .build()