+++ src/test/groovy/ru/vyarus/dropwizard/guice/cases/dwaware/DwAwareModuleApp.groovy
@@ -14,7 +14,7 @@ class DwAwareModuleApp extends Application<TestConfiguration> {
 
     @Override
     void initialize(Bootstrap<TestConfiguration> bootstrap) {
-        bootstrap.addBundle(GuiceBundle.<TestConfiguration> builder()
+        bootstrap.addBundle(GuiceBundle.builder()
                 .modules(new DwAwareModule())
                 .build()
         );