+++ src/test/groovy/ru/vyarus/dropwizard/guice/support/installerorder/OrderedInstallersApplication.groovy
@@ -20,7 +20,7 @@ class OrderedInstallersApplication extends Application<TestConfiguration> {
 
     @Override
     void initialize(Bootstrap<TestConfiguration> bootstrap) {
-        bootstrap.addBundle(GuiceBundle.<TestConfiguration> builder()
+        bootstrap.addBundle(GuiceBundle.builder()
                 .enableAutoConfig("ru.vyarus.dropwizard.guice.support.installerorder")
                 // at least one resource required
                 .extensions(DummyResource)