+++ src/test/groovy/ru/vyarus/dropwizard/guice/support/provider/InjectableProviderCheckApplication2.groovy
@@ -20,7 +20,7 @@ class InjectableProviderCheckApplication2 extends Application<TestConfiguration>
 
     @Override
     void initialize(Bootstrap<TestConfiguration> bootstrap) {
-        bootstrap.addBundle(GuiceBundle.<TestConfiguration> builder()
+        bootstrap.addBundle(GuiceBundle.builder()
                 .noDefaultInstallers()
                 .installers(JerseyProviderInstaller, ResourceInstaller)
                 .extensions(