+++ src/test/groovy/ru/vyarus/dropwizard/guice/support/provider/InjectableProviderCheckApplication3.groovy
@@ -21,7 +21,7 @@ class InjectableProviderCheckApplication3 extends Application<TestConfiguration>
 
     @Override
     void initialize(Bootstrap<TestConfiguration> bootstrap) {
-        bootstrap.addBundle(GuiceBundle.<TestConfiguration> builder()
+        bootstrap.addBundle(GuiceBundle.builder()
                 .noDefaultInstallers()
                 .installers(JerseyProviderInstaller, ResourceInstaller)
                 .extensions(