+++ src/test/groovy/ru/vyarus/dropwizard/guice/support/provider/exceptionmapper/ExceptionMapperApp2.groovy
@@ -16,7 +16,7 @@ class ExceptionMapperApp2 extends Application<TestConfiguration> {
 
     @Override
     void initialize(Bootstrap<TestConfiguration> bootstrap) {
-        bootstrap.addBundle(GuiceBundle.<TestConfiguration> builder()
+        bootstrap.addBundle(GuiceBundle.builder()
                 .noDefaultInstallers()
                 .installers(JerseyProviderInstaller, ResourceInstaller)
                 .extensions(HkManagedExceptionMapper, ExceptionResource)