+++ src/test/groovy/ru/vyarus/dropwizard/guice/support/request/RequestBeansApplication.groovy
@@ -21,7 +21,7 @@ class RequestBeansApplication extends Application<TestConfiguration> {
 
     @Override
     void initialize(Bootstrap<TestConfiguration> bootstrap) {
-        bootstrap.addBundle(GuiceBundle.<TestConfiguration> builder()
+        bootstrap.addBundle(GuiceBundle.builder()
                 .enableAutoConfig("ru.vyarus.dropwizard.guice.support.request")
                 .modules(new BindModule(RequestScopedBean))
                 .build()