+++ src/test/groovy/ru/vyarus/dropwizard/guice/support/resource/ResourceSingletonCheckApplication.groovy
@@ -18,7 +18,7 @@ class ResourceSingletonCheckApplication extends Application<TestConfiguration> {
 
     @Override
     void initialize(Bootstrap<TestConfiguration> bootstrap) {
-        bootstrap.addBundle(GuiceBundle.<TestConfiguration> builder()
+        bootstrap.addBundle(GuiceBundle.builder()
                 .enableAutoConfig("ru.vyarus.dropwizard.guice.support.resource")
                 .modules(new Module() {
                     @Override