+++ src/test/groovy/ru/vyarus/dropwizard/guice/support/web/ServletsApplication.groovy
@@ -17,7 +17,7 @@ class ServletsApplication extends Application<TestConfiguration> {
 
     @Override
     void initialize(Bootstrap<TestConfiguration> bootstrap) {
-        bootstrap.addBundle(GuiceBundle.<TestConfiguration> builder()
+        bootstrap.addBundle(GuiceBundle.builder()
                 .enableAutoConfig("ru.vyarus.dropwizard.guice.support.web.feature")
                 .modules(new WebModule(), new Module() {
                     @Override