+++ CHANGELOG.md
@@ -148,7 +148,8 @@
 * Guicey version added into BOM (dependencyManagement section in guicey pom) to avoid duplicate versions declarations
 * Java 11 compatibility. Automatic module name (in meta-inf): `dropwizard-guicey.core`
 * (breaking) AdminRestBundle moved into ext modules (bundle become guicey bundle, 
-    now return 404 instead of 403 on main context for admin-only resources)  
+    now return 404 instead of 403 on main context for admin-only resources)
+* (breaking) Remove useless configuration generic on main bundle: `GuiceBundle.<MyConfig>builder()` must be just `GuiceBundle.builder()`      
 
 Migration matrix: